# Optimize sorting calls when both keys and ranks are needed